Add 104

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,16 +9,16 @@
 int main(void)
 {
 	int i;
-	long int n1 = 1, n2 = 2, n3;
+	long long int n1 = 1, n2 = 2, n3;
 
-	printf("%ld, %ld, ", n1, n2);
+	printf("%lld, %lld, ", n1, n2);
 	for (i = 3; i <= 98; i++)
 	{
 		n3 = n1 + n2;
 		if (i == 98)
-			printf("%ld", n3);
+			printf("%lld", n3);
 		else
-			printf("%ld, ", n3);
+			printf("%lld, ", n3);
 		n1 = n2;
 		n2 = n3;
 	}
